feat: add ToolbarIconButton component
