+++ react/src/Toolbar/index.ts
@@ -1,3 +1,5 @@
 export * from './Toolbar'
 export * from './ToolbarButton'
+export * from './ToolbarDivider'
 export * from './ToolbarGroup'
+export * from './ToolbarIconButton'
